server/api/requirements.txt
python-dotenv~=0.19.2
fastapi~=0.103.2
